Create total_hacker_score
The total score of a hacker is the sum of their maximum scores for all of the challenges. Write a query to print the hacker_id, name, and total score of the hackers ordered by the descending score. If more than one hacker achieved the same total score, then sort the result by ascending hacker_id. Exclude all hackers with a total score of  from your result.


Hacker 4071 submitted solutions for challenges 19797 and 49593, so the total score .
= 95 + max(43,96)